customer ashboard redesign ongoing
