+++ src/components/CustomerProfile/dataFile.js
@@ -1,0 +1,217 @@
+const data2 = [
+  {
+    image: "",
+    title: "Personal Details",
+    subheading: "Edit Name, Email, Phone, 2FA",
+    path: "/login",
+  },
+  {
+    image: "/images/Order.png",
+    title: "Orders",
+    subheading: "Edit Name, Email, Phone, 2FA",
+    path: "/order",
+  },
+  {
+    image: "/images/Cards.png",
+    title: "Cards",
+    subheading: "Manage payment cards",
+    path: "/cards",
+  },
+  {
+    image: "/images/Message.png",
+    title: "Inbox",
+    subheading: "View messages from Azany",
+    path: "/messages",
+  },
+  {
+    image: "/images/Watchlist.png",
+    title: "Watchlist",
+    subheading: "View, modify your list",
+    path: "/watchlist",
+  },
+  {
+    image: "/images/Ratings.png",
+    title: "Ratings & Review",
+    subheading: "Edit ratings & reviews on products",
+    path: "/rating",
+  },
+  {
+    image: "/images/Track.png",
+    title: "Address",
+    subheading: "Edit your address",
+    path: "/address",
+  },
+  {
+    image: "/images/Book (2).png",
+    title: "Track Orders",
+    subheading: "Track your orders",
+    path: "/track_orders",
+  },
+];
+
+export const data = [
+  {
+    image: (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="19"
+        height="20"
+        viewBox="0 0 19 20"
+        fill="none"
+        className="text-[#8B909A] group-hover:text-white"
+      >
+        <path
+          d="M14.589 7.49618L10.5074 4.32155C9.77055 3.74829 8.73858 3.74829 8.0017 4.32155L3.91937 7.49618C3.42221 7.88281 3.13154 8.47743 3.13184 9.10723V14.6177C3.13184 15.4631 3.81715 16.1484 4.66252 16.1484H13.8466C14.692 16.1484 15.3773 15.4631 15.3773 14.6177V9.10723C15.3773 8.47735 15.0865 7.88268 14.589 7.49618"
+          stroke="currentColor"
+          stroke-width="1.46111"
+          stroke-linecap="round"
+          stroke-linejoin="round"
+        />
+        <path
+          d="M12.2929 12.3106C10.6015 13.3308 7.86004 13.3308 6.17017 12.3106"
+          stroke="currentColor"
+          stroke-width="1.46111"
+          stroke-linecap="round"
+          stroke-linejoin="round"
+        />
+      </svg>
+    ),
+    title: "Orders",
+    path: "/order",
+  },
+  {
+    image: (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="18"
+        height="18"
+        viewBox="0 0 18 18"
+        fill="none"
+        className="text-[#8B909A] group-hover:text-white"
+      >
+        <path
+          d="M15.3 7.2H16.2V17.1H1.79999V7.2H2.69999V5.4C2.69999 2.916 4.71599 0.899998 7.19999 0.899998C7.83899 0.899998 8.45099 1.035 8.99999 1.278C9.56708 1.02791 10.1802 0.899157 10.8 0.899998C13.284 0.899998 15.3 2.916 15.3 5.4V7.2ZM4.49999 5.4V7.2H6.29999V5.4C6.29999 4.383 6.65099 3.456 7.21799 2.7H7.19999C5.71499 2.7 4.49999 3.915 4.49999 5.4ZM13.5 7.2V5.4C13.5 3.915 12.285 2.7 10.8 2.7H10.782C11.3738 3.47605 11.6961 4.42406 11.7 5.4V7.2H13.5ZM8.99999 3.402C8.45099 3.897 8.09999 4.608 8.09999 5.4V7.2H9.89999V5.4C9.89999 4.608 9.54899 3.897 8.99999 3.402Z"
+          fill="currentColor"
+        />
+      </svg>
+    ),
+    title: "Addresses",
+    path: "/address",
+  },
+  {
+    image: (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="18"
+        height="18"
+        viewBox="0 0 18 18"
+        fill="none"
+        className="text-[#8B909A] group-hover:text-white"
+      >
+        <path
+          d="M9.225 9.16488C9.62518 8.81848 9.94616 8.39004 10.1662 7.90865C10.3861 7.42725 10.5 6.90416 10.5 6.37488C10.5 5.38032 10.1049 4.42649 9.40165 3.72323C8.69839 3.01997 7.74456 2.62488 6.75 2.62488C5.75544 2.62488 4.80161 3.01997 4.09835 3.72323C3.39509 4.42649 3 5.38032 3 6.37488C2.99999 6.90416 3.11385 7.42725 3.33384 7.90865C3.55384 8.39004 3.87482 8.81848 4.275 9.16488C3.22511 9.64029 2.33435 10.408 1.70924 11.3763C1.08412 12.3445 0.751104 13.4724 0.75 14.6249C0.75 14.8238 0.829018 15.0146 0.96967 15.1552C1.11032 15.2959 1.30109 15.3749 1.5 15.3749C1.69891 15.3749 1.88968 15.2959 2.03033 15.1552C2.17098 15.0146 2.25 14.8238 2.25 14.6249C2.25 13.4314 2.72411 12.2868 3.56802 11.4429C4.41193 10.599 5.55653 10.1249 6.75 10.1249C7.94347 10.1249 9.08807 10.599 9.93198 11.4429C10.7759 12.2868 11.25 13.4314 11.25 14.6249C11.25 14.8238 11.329 15.0146 11.4697 15.1552C11.6103 15.2959 11.8011 15.3749 12 15.3749C12.1989 15.3749 12.3897 15.2959 12.5303 15.1552C12.671 15.0146 12.75 14.8238 12.75 14.6249C12.7489 13.4724 12.4159 12.3445 11.7908 11.3763C11.1657 10.408 10.2749 9.64029 9.225 9.16488ZM6.75 8.62488C6.30499 8.62488 5.86998 8.49292 5.49997 8.24568C5.12996 7.99845 4.84157 7.64705 4.67127 7.23592C4.50097 6.82478 4.45642 6.37238 4.54323 5.93592C4.63005 5.49947 4.84434 5.09856 5.15901 4.78389C5.47368 4.46922 5.87459 4.25493 6.31105 4.16811C6.7475 4.08129 7.1999 4.12585 7.61104 4.29615C8.02217 4.46645 8.37357 4.75483 8.62081 5.12485C8.86804 5.49486 9 5.92987 9 6.37488C9 6.97161 8.76295 7.54391 8.34099 7.96587C7.91903 8.38783 7.34674 8.62488 6.75 8.62488ZM14.055 8.86488C14.535 8.32438 14.8485 7.65667 14.9579 6.94213C15.0672 6.2276 14.9677 5.49668 14.6713 4.83738C14.375 4.17807 13.8943 3.61847 13.2874 3.22594C12.6804 2.83342 11.9729 2.62468 11.25 2.62488C11.0511 2.62488 10.8603 2.7039 10.7197 2.84455C10.579 2.9852 10.5 3.17597 10.5 3.37488C10.5 3.57379 10.579 3.76456 10.7197 3.90521C10.8603 4.04586 11.0511 4.12488 11.25 4.12488C11.8467 4.12488 12.419 4.36193 12.841 4.78389C13.2629 5.20584 13.5 5.77814 13.5 6.37488C13.4989 6.76881 13.3945 7.15556 13.197 7.49644C12.9996 7.83733 12.7162 8.12041 12.375 8.31738C12.2638 8.38152 12.1709 8.47313 12.1053 8.58343C12.0396 8.69374 12.0034 8.81905 12 8.94738C11.9969 9.0747 12.0262 9.20073 12.0852 9.31357C12.1443 9.42642 12.2311 9.52237 12.3375 9.59238L12.63 9.78738L12.7275 9.83988C13.6315 10.2687 14.3942 10.9469 14.9257 11.7946C15.4572 12.6424 15.7354 13.6243 15.7275 14.6249C15.7275 14.8238 15.8065 15.0146 15.9472 15.1552C16.0878 15.2959 16.2786 15.3749 16.4775 15.3749C16.6764 15.3749 16.8672 15.2959 17.0078 15.1552C17.1485 15.0146 17.2275 14.8238 17.2275 14.6249C17.2336 13.4739 16.9454 12.3406 16.3901 11.3324C15.8348 10.3243 15.031 9.47486 14.055 8.86488Z"
+          fill="currentColor"
+        />
+      </svg>
+    ),
+    title: "Recently Viewed",
+    path: "/recently-viewed",
+  },
+  {
+    image: (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="19"
+        height="19"
+        viewBox="0 0 19 19"
+        fill="none"
+        className="text-[#8B909A] group-hover:text-white"
+      >
+        <g clip-path="url(#clip0_263_6704)">
+          <path
+            d="M8.48505 1.66528L15.7975 5.32153V14.4622L8.48505 18.1096L1.17255 14.4622V5.32153L8.48505 1.66528ZM13.9782 5.6731L8.48505 2.93091L6.36688 3.9856L11.8249 6.74536L13.9782 5.6731ZM8.48505 8.41528L10.5768 7.37817L5.11005 4.61841L2.99188 5.6731L8.48505 8.41528ZM2.29755 6.58716V13.759L7.92255 16.5715V9.39966L2.29755 6.58716ZM9.04755 16.5715L14.6725 13.759V6.58716L9.04755 9.39966V16.5715Z"
+            fill="currentColor"
+          />
+        </g>
+        <defs>
+          <clipPath id="clip0_263_6704">
+            <rect
+              width="18"
+              height="18"
+              fill="white"
+              transform="translate(0.0475464 0.610596)"
+            />
+          </clipPath>
+        </defs>
+      </svg>
+    ),
+    title: "Saved Items",
+    path: "/saved-items",
+  },
+
+  {
+    image: (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="19"
+        height="19"
+        viewBox="0 0 19 19"
+        fill="none"
+        className="text-[#8B909A] group-hover:text-white"
+      >
+        <path
+          d="M11.2975 5.27637C11.085 5.27637 10.9068 5.20437 10.7628 5.06037C10.6188 4.91637 10.547 4.73837 10.5475 4.52637C10.5475 4.31387 10.6195 4.13562 10.7635 3.99162C10.9075 3.84762 11.0855 3.77587 11.2975 3.77637H15.7975C16.01 3.77637 16.1883 3.84837 16.3323 3.99237C16.4763 4.13637 16.548 4.31437 16.5475 4.52637C16.5475 4.73887 16.4755 4.91712 16.3315 5.06112C16.1875 5.20512 16.0095 5.27687 15.7975 5.27637H11.2975ZM11.2975 8.27637C11.085 8.27637 10.9068 8.20437 10.7628 8.06037C10.6188 7.91637 10.547 7.73837 10.5475 7.52637C10.5475 7.31387 10.6195 7.13562 10.7635 6.99162C10.9075 6.84762 11.0855 6.77587 11.2975 6.77637H15.7975C16.01 6.77637 16.1883 6.84837 16.3323 6.99237C16.4763 7.13637 16.548 7.31437 16.5475 7.52637C16.5475 7.73887 16.4755 7.91712 16.3315 8.06112C16.1875 8.20512 16.0095 8.27687 15.7975 8.27637H11.2975ZM11.2975 11.2764C11.085 11.2764 10.9068 11.2044 10.7628 11.0604C10.6188 10.9164 10.547 10.7384 10.5475 10.5264C10.5475 10.3139 10.6195 10.1356 10.7635 9.99162C10.9075 9.84762 11.0855 9.77587 11.2975 9.77637H15.7975C16.01 9.77637 16.1883 9.84837 16.3323 9.99237C16.4763 10.1364 16.548 10.3144 16.5475 10.5264C16.5475 10.7389 16.4755 10.9171 16.3315 11.0611C16.1875 11.2051 16.0095 11.2769 15.7975 11.2764H11.2975ZM6.04755 10.5264C5.42255 10.5264 4.8913 10.3076 4.4538 9.87012C4.0163 9.43262 3.79755 8.90137 3.79755 8.27637C3.79755 7.65137 4.0163 7.12012 4.4538 6.68262C4.8913 6.24512 5.42255 6.02637 6.04755 6.02637C6.67255 6.02637 7.2038 6.24512 7.6413 6.68262C8.0788 7.12012 8.29755 7.65137 8.29755 8.27637C8.29755 8.90137 8.0788 9.43262 7.6413 9.87012C7.2038 10.3076 6.67255 10.5264 6.04755 10.5264ZM1.54755 14.2764V13.6014C1.54755 13.3389 1.61005 13.0889 1.73505 12.8514C1.86005 12.6139 2.03505 12.4264 2.26005 12.2889C2.82255 11.9514 3.41955 11.6984 4.05105 11.5299C4.68255 11.3614 5.34805 11.2769 6.04755 11.2764C6.74755 11.2764 7.4133 11.3609 8.0448 11.5299C8.6763 11.6989 9.27305 11.9519 9.83505 12.2889C10.06 12.4264 10.235 12.6139 10.36 12.8514C10.485 13.0889 10.5475 13.3389 10.5475 13.6014V14.2764C10.5475 14.4889 10.4755 14.6671 10.3315 14.8111C10.1875 14.9551 10.0095 15.0269 9.79755 15.0264H2.29755C2.08505 15.0264 1.9068 14.9544 1.7628 14.8104C1.6188 14.6664 1.54705 14.4884 1.54755 14.2764ZM3.16005 13.5264H8.93505C8.49755 13.2764 8.03505 13.0889 7.54755 12.9639C7.06005 12.8389 6.56005 12.7764 6.04755 12.7764C5.53505 12.7764 5.03505 12.8389 4.54755 12.9639C4.06005 13.0889 3.59755 13.2764 3.16005 13.5264ZM6.04755 9.02637C6.26005 9.02637 6.4383 8.95437 6.5823 8.81037C6.7263 8.66637 6.79805 8.48837 6.79755 8.27637C6.79755 8.06387 6.72555 7.88562 6.58155 7.74162C6.43755 7.59762 6.25955 7.52587 6.04755 7.52637C5.83505 7.52637 5.6568 7.59837 5.5128 7.74237C5.3688 7.88637 5.29705 8.06437 5.29755 8.27637C5.29755 8.48887 5.36955 8.66712 5.51355 8.81112C5.65755 8.95512 5.83555 9.02687 6.04755 9.02637Z"
+          fill="currentColor"
+        />
+      </svg>
+    ),
+    title: "Reviews",
+    path: "/reviews",
+  },
+  {
+    image: (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="19"
+        height="19"
+        viewBox="0 0 19 19"
+        fill="none"
+        className="text-[#8B909A] group-hover:text-white"
+      >
+        <path
+          d="M5.87762 15.6955C6.85329 16.2152 7.9421 16.4861 9.04755 16.4843C12.7755 16.4843 15.7975 13.4622 15.7975 9.73431C15.7975 6.00638 12.7755 2.98431 9.04755 2.98431C5.31962 2.98431 2.29755 6.00638 2.29755 9.73431C2.29755 10.8818 2.58362 11.9612 3.08762 12.9062C3.20521 13.1263 3.23286 13.3835 3.16476 13.6236L2.37147 16.4117L5.16019 15.619C5.40022 15.5507 5.65738 15.5781 5.87762 15.6955ZM9.04755 17.4486C7.78376 17.4504 6.53906 17.1404 5.42376 16.546L2.35733 17.4184C2.21968 17.4575 2.07408 17.4591 1.93559 17.4231C1.79709 17.3871 1.67072 17.3148 1.56954 17.2136C1.46835 17.1124 1.39602 16.986 1.36002 16.8475C1.32402 16.7091 1.32566 16.5635 1.36476 16.4258L2.23647 13.3594C1.6417 12.2438 1.33145 10.9986 1.33326 9.73431C1.33326 5.47409 4.78733 2.02002 9.04755 2.02002C13.3078 2.02002 16.7618 5.47409 16.7618 9.73431C16.7618 13.9945 13.3078 17.4486 9.04755 17.4486ZM6.15469 8.28788C6.15469 8.16 6.20549 8.03737 6.29591 7.94695C6.38632 7.85653 6.50896 7.80573 6.63683 7.80573H11.4583C11.5861 7.80573 11.7088 7.85653 11.7992 7.94695C11.8896 8.03737 11.9404 8.16 11.9404 8.28788C11.9404 8.41575 11.8896 8.53838 11.7992 8.6288C11.7088 8.71922 11.5861 8.77002 11.4583 8.77002H6.63683C6.50896 8.77002 6.38632 8.71922 6.29591 8.6288C6.20549 8.53838 6.15469 8.41575 6.15469 8.28788ZM6.15469 11.1807C6.15469 11.0529 6.20549 10.9302 6.29591 10.8398C6.38632 10.7494 6.50896 10.6986 6.63683 10.6986H9.52969C9.65756 10.6986 9.7802 10.7494 9.87062 10.8398C9.96103 10.9302 10.0118 11.0529 10.0118 11.1807C10.0118 11.3086 9.96103 11.4312 9.87062 11.5217C9.7802 11.6121 9.65756 11.6629 9.52969 11.6629H6.63683C6.50896 11.6629 6.38632 11.6121 6.29591 11.5217C6.20549 11.4312 6.15469 11.3086 6.15469 11.1807Z"
+          fill="currentColor"
+        />
+      </svg>
+    ),
+    title: "Messages",
+    path: "/messages",
+  },
+  {
+    image: (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="18"
+        height="18"
+        viewBox="0 0 18 18"
+        fill="none"
+        className="text-[#8B909A] group-hover:text-white"
+      >
+        <path
+          fill-rule="evenodd"
+          clip-rule="evenodd"
+          d="M15.825 6.00495C15.7837 6.00195 15.738 6.00195 15.69 6.00195H13.7962C12.2452 6.00195 10.9185 7.22295 10.9185 8.81445C10.9185 10.406 12.246 11.627 13.7962 11.627H15.69C15.738 11.627 15.7837 11.627 15.8265 11.624C16.1453 11.6047 16.4462 11.4703 16.6732 11.2456C16.9001 11.021 17.0377 10.7215 17.0602 10.403C17.0632 10.358 17.0632 10.3092 17.0632 10.2642V7.3647C17.0632 7.3197 17.0632 7.27095 17.0602 7.22595C17.0377 6.9074 16.9001 6.6079 16.6732 6.38327C16.4462 6.15863 16.1453 6.02417 15.8265 6.00495H15.825ZM13.629 9.56445C14.028 9.56445 14.3512 9.22845 14.3512 8.81445C14.3512 8.40045 14.028 8.06445 13.629 8.06445C13.2292 8.06445 12.906 8.40045 12.906 8.81445C12.906 9.22845 13.2292 9.56445 13.629 9.56445Z"
+          fill="currentColor"
+        />
+        <path
+          fill-rule="evenodd"
+          clip-rule="evenodd"
+          d="M15.6885 12.752C15.7144 12.7509 15.7402 12.756 15.7638 12.7668C15.7874 12.7776 15.8081 12.7938 15.8243 12.8141C15.8404 12.8345 15.8516 12.8583 15.8568 12.8837C15.862 12.9091 15.8611 12.9354 15.8542 12.9605C15.7043 13.4945 15.465 13.9505 15.0817 14.333C14.52 14.8955 13.8083 15.1437 12.9293 15.2622C12.0743 15.377 10.983 15.377 9.6045 15.377H8.0205C6.642 15.377 5.55 15.377 4.69575 15.2622C3.81675 15.1437 3.105 14.8947 2.54325 14.3337C1.98225 13.772 1.73325 13.0602 1.61475 12.1812C1.5 11.3262 1.5 10.235 1.5 8.85645V8.77245C1.5 7.39395 1.5 6.30195 1.61475 5.44695C1.73325 4.56795 1.98225 3.8562 2.54325 3.29445C3.105 2.73345 3.81675 2.48445 4.69575 2.36595C5.55075 2.25195 6.642 2.25195 8.0205 2.25195H9.6045C10.983 2.25195 12.075 2.25195 12.9293 2.3667C13.8083 2.4852 14.52 2.7342 15.0817 3.2952C15.465 3.6792 15.7043 4.13445 15.8542 4.66845C15.8611 4.69348 15.862 4.71977 15.8568 4.74519C15.8516 4.77062 15.8404 4.79445 15.8243 4.81476C15.8081 4.83508 15.7874 4.85129 15.7638 4.8621C15.7402 4.87291 15.7144 4.87799 15.6885 4.87695H13.7955C11.6677 4.87695 9.79275 6.55695 9.79275 8.81445C9.79275 11.072 11.6677 12.752 13.7955 12.752H15.6885ZM4.3125 5.25195C4.16332 5.25195 4.02024 5.31122 3.91475 5.41671C3.80926 5.52219 3.75 5.66527 3.75 5.81445C3.75 5.96364 3.80926 6.10671 3.91475 6.2122C4.02024 6.31769 4.16332 6.37695 4.3125 6.37695H7.3125C7.46168 6.37695 7.60476 6.31769 7.71025 6.2122C7.81574 6.10671 7.875 5.96364 7.875 5.81445C7.875 5.66527 7.81574 5.52219 7.71025 5.41671C7.60476 5.31122 7.46168 5.25195 7.3125 5.25195H4.3125Z"
+          fill="currentColor"
+        />
+      </svg>
+    ),
+    title: "Reward(coming soon)",
+    path: "/rewards",
+  },
+];
